added functions for handle img
