+++ .vscode/settings.json
@@ -1,17 +1,22 @@
 {
-    "files.associations": {
-		"float_vector.h": "c",
-		"libft.h": "c",
-		"mlx_int.h": "c",
-		"mlx.h": "c",
-		"x.h": "c",
-		"keysym.h": "c",
-		"keysymdef.h": "c",
-		"cstdlib": "c",
-		"array": "c",
-		"string": "c",
-		"string_view": "c",
-		"cerrno": "c",
-		"cub3d.h": "c"
-	}
+  "files.associations": {
+    "*.excalidraw": "json",
+    "*.excalidrawlib": "json",
+    "float_vector.h": "c",
+    "libft.h": "c",
+    "mlx_int.h": "c",
+    "mlx.h": "c",
+    "x.h": "c",
+    "keysym.h": "c",
+    "keysymdef.h": "c",
+    "cstdlib": "c",
+    "array": "c",
+    "string": "c",
+    "string_view": "c",
+    "cerrno": "c",
+    "cub3d.h": "c",
+    "prototype.h": "c",
+    "struct.h": "c",
+    "errno.h": "c"
+  }
 }
