+++ include/define.h
@@ -1,0 +1,20 @@
+# ifndef DEFINE_H
+# define DEFINE_H
+
+# define A_KEY_CONST 'a'
+# define S_KEY_CONST 's'
+# define D_KEY_CONST 'd'
+# define W_KEY_CONST 'w'
+# define ESC_KEY_CONST 65307
+
+# define ERR_ARGS "Invalid call. Must be: ./cub3D <map_path/map.cub>"
+# define ERR_MALC "Problems in memory allocation!"
+# define ERR_CUB "Error File. Expected .cub extension!"
+# define ERR_MAP "Invalid map!"
+
+# define HEIGHT 600
+# define WIDTH 800
+# define TITLE "Cub3D"
+
+
+# endif
