+++ include/enum.h
@@ -1,0 +1,13 @@
+# ifndef ENUM_H
+# define ENUM_H
+
+enum	e_output
+{
+	SUCCESS = 0,
+	FAILURE = 1,
+	// ERROR = 2,
+	BREAK = 3,
+	CONTINUE = 4
+};
+
+# endif
