+++ lib/libft/src/ft_get_next_line.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_get_next_line.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/16 07:05:38 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/14 11:27:48 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:56:17 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 static int	is_endl(char *str)
 {
