+++ lib/libft/src/ft_print_pointer.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_print_pointer.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 03:59:54 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/20 17:05:49 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:55:53 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 int	ft_print_pointer(void *ptr)
 {
