+++ lib/libft/src/ft_printf.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 03:52:45 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/20 17:08:55 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:55:50 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 static int	ft_manage_specifier(va_list arg, char specifier)
 {
