+++ lib/libft/src/ft_putchar.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putchar.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 03:47:53 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/13 13:50:49 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:55:47 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 void	ft_putchar(char c)
 {
