+++ lib/libft/src/ft_putstr.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_putstr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/17 03:51:08 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/20 17:04:05 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:55:44 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 int	ft_putstr(char *str)
 {
