+++ lib/libft/src/ft_split.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_split.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rlins <rlins@student.42.fr>                +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/14 10:46:59 by rlins             #+#    #+#             */
-/*   Updated: 2022/08/14 10:47:20 by rlins            ###   ########.fr       */
+/*   Updated: 2023/03/11 00:55:42 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../../include/libft.h"
+#include "libft.h"
 
 static int	str_counter(char *str, char c)
 {
