+++ src/graph/mlx_color.c
@@ -1,0 +1,7 @@
+#include "cub3d.h"
+
+int ft_mlx_rgb_to_int(int r, int g, int b)
+{
+    return (r << 16 | g << 8 | b);
+}
+
