+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/04 15:22:49 by rlins             #+#    #+#             */
-/*   Updated: 2023/03/10 23:14:48 by lucas            ###   ########.fr       */
+/*   Updated: 2023/03/11 01:47:14 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,6 @@
 
 static int	args_handler(t_data *data, char **argv);
 
-void	init_data(t_data *data)
-{
-	ft_bzero(data, sizeof(t_data));
-	data->player.pos_x = 22;
-	data->player.pos_y = 12;
-	data->player.dir_x = -1;
-	data->player.plane_y = 0.66;
-}
 
 int	main(int argc, char **argv)
 {
@@ -30,7 +22,7 @@ int	main(int argc, char **argv)
 	if (argc != 2)
 		return (error_msg(ERR_ARGS, 1));
 	init_data(&data);
-	create_graphic(&data);
+	ft_mlx_init(&data);
 	if (args_handler(&data, argv) != 0)
 		return (EXIT_FAILURE);
 	mlx_loop(data.view.mlx);
